% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{Aggr_step}
\alias{Aggr_step}
\title{Step4: Aggregation}
\usage{
Aggr_step(in.T1D1_S2, in.T2D2_S2, in.id.S2, para.q)
}
\arguments{
\item{in.T1D1_S2}{numeric vector, contains \eqn{T_1^{(k)}} for all \eqn{k} in \eqn{\mathcal{S}_2}.}

\item{in.T2D2_S2}{numeric vector, contains \eqn{T_2^{(k)}} for all \eqn{k} in \eqn{\mathcal{S}_2}.}

\item{in.id.S2}{vector, indices of the mediators retained by the penalized Cox model, that is \eqn{\mathcal{S}_2}.}

\item{para.q}{numeric value from 0 to 1. The target FDR level.}
}
\value{
\code{id.S3}, the finally selected mediators, that is \eqn{\hat{\mathcal{H}}_1}.
}
\description{
Aggregate the \eqn{T_1} and \eqn{T_2} derived in step 3 for inference. First construct \eqn{W_k} and calculate the threshold \eqn{L}. Then select
\deqn{\hat{\mathcal{H}}_1 = \{k : W_k \ge L\}.}
}
