% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{Deran_step}
\alias{Deran_step}
\title{Derandomization Operation}
\usage{
Deran_step(in.S3list, para.freq = 0.5)
}
\arguments{
\item{in.S3list}{list, with \code{B} elements. \code{b}-th element is the set of selected mediators derived from the \code{b}-th sample splitting.}

\item{freq}{numeric value from \code{0} to \code{1}, default as \code{0.5}. The frequency threshold which is used to determine whether an index should be included in the voting set.}
}
\value{
A list with elements \code{id.S3.deran} , \code{B} , \code{freq} , \code{S_vote} , \code{bstar} :

\code{id.S3.deran}  the indices of mediators selected by derandomization operation.

\code{B}  the total times of sample splitting.

\code{freq}  a table which is used to illustrate the frequency of each mediator present in results of multiple sample splitting.

\code{S_vote}  the voting set.

\code{bstar}  the indice of which set has the max overlap with the voting set.
}
\description{
The function to aggregate the results from multiple sample splitting.
First we generate the voting set and choose the set which has max overlap with voting set.
Voting set is defined as
\deqn{ \hat{\mathcal{H}}_1^{(vote)} = \left\{ k:\sum_{b = 1}^{B}{ I(k \in \hat{\mathcal{H}}_1^{(b)}) } \ge B \times freq \right\}.}
}
