% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get.MISAthreshold}
\alias{get.MISAthreshold}
\title{Threshold Calculation}
\usage{
get.MISAthreshold(Wk, g.q)
}
\arguments{
\item{Wk}{vector, the value of mirror statistics \eqn{W_k}.}

\item{g.q}{numeric, from 0 to 1. The target FDR level.}
}
\value{
A threshold value \eqn{L} according to specified target FDR level \eqn{q}.
}
\description{
A function to calculate the threshold \eqn{L} for Mirror statistics.
}
\section{Details}{

The threshold \eqn{L} is calculated as follow:
\deqn{L = \inf \left\{ t: \frac{ |\{k: W_k \le -t\}| }{ \max(|\{k: W_k \ge t\}|,1) } \le q , k \in \mathcal{S}_2 \right\}}
}

