% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{PreScr_step}
\alias{PreScr_step}
\title{Step1: Pre-screening by SIS procedure}
\usage{
PreScr_step(
  in.X,
  in.Z,
  in.M,
  para.kappa = 1,
  para.parallel = F,
  para.coresnum,
  para.d,
  in.stat_prescr
)
}
\arguments{
\item{in.X}{vector, with \code{n} elements or matirx with \code{1} column, the input of one-dimensional exposures.}

\item{in.Z}{matrix, with \code{n} rows and \code{s} colnumns, the input of s-dimensional covariates.}

\item{in.M}{matrix, with \code{n} rows and \code{p} colnumns, the input of p-dimensional candidate mediators.}

\item{para.kappa}{numeric, suggested to be \code{1} or \code{2}. When \code{para.d} is not specified, it will be used to calculate \code{para.d} as \eqn{d = n/log(n) * \kappa}. We recommended to determine \code{para.d} through \code{para.kappa} rather than manually. Refer to \code{para.d}.}

\item{para.parallel}{logical, default as \code{FALSE}. Determines whether to use parallel computation for acceleration.}

\item{para.coresnum}{integer, the number of cores to use when \code{para.parallel} is set to \code{TRUE}. Default as all the cores detected by \code{detectCores()}.}

\item{para.d}{integer, the number of mediators will be retained by SIS program. If \code{para.d} is not specified, we will use \code{para.kappa} and the sample size \code{n} to calculate it. Refer to \code{para.kappa}.}

\item{in.stat_prescr}{vector with length p, the user-specified filter indicator. The larger absolute value, the more likely to be retained. When it is provided, the marginal effect of X-M will no longer be calculated.}
}
\value{
vector with length \code{para.d}, a set of indices containing the mediators retained by the SIS program.
}
\description{
The first step of survMISA involves pre-screening candidate mediators to a moderate level using the SIS procedure. While we recommend utilizing the X-M side information, you may also specify your preferred indicator. In cases where the number of candidate variables is extremely high, the pre-screening step can significantly impact the final result, so careful parameter selection is essential.
}
