% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{SampleSplit_step}
\alias{SampleSplit_step}
\title{Step2: sample splitting}
\usage{
SampleSplit_step(in.n, para.nu = 1, para.seed = 1)
}
\arguments{
\item{in.n}{integer, the size of whole sample.}

\item{para.nu}{numeric, to calculate the size of subsample as:
\deqn{n_1 : n_2 = \nu.}}

\item{para.seed}{numeric, random seed for random splitting. Use to fix the result.}
}
\value{
a list containing two sets, \code{indv1} and \code{indv2}, which represent the indices of the two subsamples, respectively.
}
\description{
A function designed to randomly divide the sample of size n into two complementary subsets, with sizes \eqn{n_1} and \eqn{n_2}, respectively.
}
\examples{
\code{SampleSplit_step(100,1,1)}
\code{SampleSplit_step(100,2,1)}
SampleSplit_step(100,2,2)
}
