% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{T1onD1}
\alias{T1onD1}
\title{Step3.1: Calculate \eqn{T_1^{(k)}} on \eqn{\mathcal{D}_1}}
\usage{
T1onD1(
  in.X,
  in.Z,
  in.M,
  in.Y,
  in.id.S1,
  in.indv1,
  para.rs = c("half", "BIC", "1se"),
  para.parallel = F,
  para.coresnum
)
}
\arguments{
\item{in.X}{vector, with \code{n} elements or matirx with \code{1} column, the input of one-dimensional exposures.}

\item{in.Z}{matrix, with \code{n} rows and \code{s} colnumns, the input of s-dimensional covariates.}

\item{in.M}{matrix, with \code{n} rows and \code{p} colnumns, the input of p-dimensional candidate mediators.}

\item{in.Y}{\code{Surv} object, with \code{n} elements, the input of survival information.}

\item{in.id.S1}{vector, indices of the mediators retained by the SIS program.}

\item{in.indv1}{vector, indices of the individuals in the first subsample \eqn{\mathcal{D}_1}}

\item{para.rs}{one of \code{half}, \code{BIC} and \verb{1se}. The strategy for selecting the penalty parameter \eqn{\lambda} in the penalized Cox model. The \code{half} option retains half of the mediators in \eqn{\mathcal{S}_1}. The \code{BIC} and \verb{1se} options represent \code{lambda.min} and \code{lambda.1se} from the \code{cv.glmnet} function, respectively.}

\item{para.parallel}{logical, default as \code{FALSE}. Determines whether to use parallel computation for acceleration.}

\item{para.coresnum}{integer, the number of cores to use when \code{para.parallel} is set to \code{TRUE}. Default as all the cores detected by \code{detectCores()}.}
}
\value{
a list with elements \code{T1} , \code{T1_part} , \code{id.S1} , \code{id.S2} , \code{indv} :

\code{T1} is \eqn{T_1^{(k)},\forall k \in \mathcal{S}_1}.

\code{T1_part} is \eqn{T_1^{(k)},\forall k \in \mathcal{S}_2}.

\code{id.S1} and \code{id.S2} contains the indices of sets \eqn{\mathcal{S}_1} and \eqn{\mathcal{S}_2}, respectively.

\code{indv} contains the indices in subsample \eqn{\mathcal{D}_1}.
}
\description{
In step3.1, we will calculate \eqn{T_1^{(k)}} on \eqn{\mathcal{D}_1}. We first fit partially penalized Cox model on \eqn{\mathcal{D}_1} and \eqn{\mathcal{S}_1} for further screening and denote the retained mediators as \eqn{\mathcal{S}_2}. Then calculate \eqn{\tilde{\beta}_k} and \eqn{\hat{\alpha}_k^{(1)}} on \eqn{\mathcal{S}_2}. Finally calculate
\deqn{T_1^{(k)} = \begin{cases} \frac{ \hat{\alpha}_k^{(1)} \tilde{\beta}_k}{\hat{se}_{\alpha,k}^{(1)}}, & k \in \mathcal{S}_2 \\ 0, & \mbox{else}\end{cases}.}
}
