% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{T2onD2}
\alias{T2onD2}
\title{Step3.2: Calculate \eqn{T_2^{(k)}} on \eqn{\mathcal{D}_2}}
\usage{
T2onD2(
  in.X,
  in.Z,
  in.M,
  in.Y,
  in.id.S2,
  in.indv2,
  para.parallel = F,
  para.coresnum
)
}
\arguments{
\item{in.X}{vector, with \code{n} elements or matirx with \code{1} column, the input of one-dimensional exposures.}

\item{in.Z}{matrix, with \code{n} rows and \code{s} colnumns, the input of s-dimensional covariates.}

\item{in.M}{matrix, with \code{n} rows and \code{p} colnumns, the input of p-dimensional candidate mediators.}

\item{in.Y}{\code{Surv} object, with \code{n} elements, the input of survival information.}

\item{in.id.S2}{vector, indices of the mediators retained by the penalized Cox model.}

\item{in.indv2}{vector, indices of the individuals in the first subsample \eqn{\mathcal{D}_2}}

\item{para.parallel}{logical, default as \code{FALSE}. Determines whether to use parallel computation for acceleration.}

\item{para.coresnum}{integer, the number of cores to use when \code{para.parallel} is set to \code{TRUE}. Default as all the cores detected by \code{detectCores()}.}
}
\value{
a list with elements \code{T2_part} , \code{id.S2} , \code{indv2} :

\code{T2_part} is \eqn{T_2^{(k)},\forall k \in \mathcal{S}_2}.

\code{id.S2} contains the indices of the set \eqn{\mathcal{S}_2}.

\code{indv2} contains the indices in subsample \eqn{\mathcal{D}_2}.
}
\description{
In step3.2, we will calculate \eqn{T_2^{(k)}} on \eqn{\mathcal{D}_2}.
On \eqn{\mathcal{D}_2}, we obtain \eqn{\hat{\alpha}_k^{(2)}}, \eqn{\hat{\beta}_k^{(2)}}, \eqn{\hat{se}_{\alpha,k}^{(2)}} and \eqn{\hat{se}_{\beta,k}^{(2)}}, with selected mediators in \eqn{\mathcal{S}_2}. Here \eqn{\hat{\alpha}_k^{(2)}} and \eqn{\hat{\beta}_k^{(2)}} are the OLS and the maximum partial likelihood estimators of \eqn{\alpha_k} and \eqn{\beta_k}, respectively. Then we construct \eqn{T_2} on \eqn{\mathcal{D}_2} as follow:
\deqn{ T_2^{(k)}= \begin{cases} \frac{ \hat{\alpha}_k^{(2)} \hat{\beta}_k^{(2)} }{ \hat{se}_{\alpha,k}^{(2)} \hat{se}_{\beta,k}^{(2)} }, & k \in \mathcal{S}_2 \\ 0, & \mbox{else} \end{cases}.}
}
